B-AIA-210-LYN-2-1-eliza-9-main/requirements.txt
flask==3.0.3
httpx==0.27.2
flask-cors==3.0.10
gunicorn==20.1.0
pytest==7.3.1
pytest-cov==4.0.0
coverage==7.2.7
discord.py>=2.0